Tissue Compartments

This page introcudes Compartment types with fields of relevant tissue parameters and forward functions inferencing signals from the compartment model and a imaging protocol. Featuring spherical mean based models with compartmental relaxation-weighting.

Overview

Microstructure.Compartment — Type

Compartment Type is an abstract type that includes the Cylinder, Stick, Zeppelin, Sphere and Iso type. A Compartment Type object contains relevant tissue parameters that affect the MRI signals. Each type of compartment contain a 't2' field for combined-diffusion-relaxometry imaging. When your data supports only T2-weighted compartment modelling, i.e. acquired with single-TE, set the 't2' field to zero to for conventional dMRI modelling.

Axonal and dendritic compartments

Microstructure.Cylinder — Type

Cylinder(
    da::Float64, 
    dpara::Float64, 
    d0::Float64, 
    t2::Float64
    )

Return a Cylinder Type object with the cylinder diameter 'da', parallel diffusivity 'dpara', the intrinsic diffusivity 'd0' and the T2 relaxation time 't2'.

Examples

julia> Cylinder(da = 3.0e-6, dpara = 1.8e-9, d0 = 1.7e-9, t2 = 90e-3)
Cylinder(3.0e-6, 1.8e-9, 1.7e-9, 0.09)

Microstructure.Stick — Type

Stick(dpara::Float64, t2::Float64)

Return a Stick Type object with parallel diffusivity 'dpara' and T2 relaxation time 't2'. The perpendicular diffusivity of a Stick model is zero.

Examples

julia> Stick(dpara = 1.7e-6, t2 = 60e-3)
Stick(1.7e-6, 0.06)

Extra-cellular compartment

Microstructure.Zeppelin — Type

Zeppelin(
    dpara::Float64, 
    dperp_frac::Float64, 
    t2::Float64
    )

Return a Zeppelin Type object with parallel diffusivity 'dpara', axially symmetric perpendicular diffusivity represented as a fraction of the parallel diffusivity 'dperp_frac', and the T2 relaxation time 't2'.

Examples

julia> Zeppelin(dpara = 1.7e-6, dperp_frac = 0.5, t2 = 0.0)
Zeppelin(1.7e-6, 0.5, 0.0)

Cell body compartment

Microstructure.Sphere — Type

Sphere(
    diff::Float64, 
    size::Float64, 
    t2::Float64
    )

Return a Sphere Type object with diffusivity within sphere 'diff', spherical radius 'size', and T2 relaxation time 't2'.

Examples

julia> Sphere(diff = 3.0e-9, size = 8.0e-6, t2 = 45e-3)
Sphere(3.0e-9, 8.0e-6, 0.045)

CSF and dot compartment

Microstructure.Iso — Type

Iso(diff::Float64, t2=Float64)

Return an isotropic tensor with diffusivity 'diff' and T2 relaxation time 't2'. This compartment can be used to represent CSF (diff = free water) or dot compartment (diff = 0). The latter is for immobile water typically seen in ex vivo tissue.

Examples

julia> Iso(diff = 3.0e-9,t2 = 2000.0e-3)
Iso(3.0e-9, 2.0)

julia> Iso(diff = 0.0)
Iso(0.0, 0.0)

Compartment signals

This function implements different methods for different compartment Types to generate compartment signals.

Microstructure.compartment_signals — Function

compartment_signals(model::Compartment,protocol::Protocol)

Return compartment signals given a compartment object 'model' and a imaging 'protocol'. 'model' can be the Cylinder/Zeppelin/Stick/Sphere/Iso Type. When t2 in compartment 'model' is set as default (0), relaxation-weightings are not considered.

Citations

Constructing...