Microstructure Models

dMRI models

WM models

Microstructure.ExCaliber — Type

ExCaliber(axon, extra, csf, dot, fracs)

ExCaliber model for ex vivo tissue; dot signal considered The fraction vector represents fractions of the axon, CSF and dot with the fraction of extra being 1-sum(fracs)

GM models

Microstructure.SANDI — Type

SANDI(soma, neurite, extra, fracs)

SANDI models or MTE-SANDI models including three tissue compartments for in vivo imaging For SANDI model, ignore the field of t2 in all compartments and set them to 0

Microstructure.SANDIdot — Type

SANDIdot models or MTE-SANDIdot models including additionally a dot compartment for ex vivo imaging For SANDIdot model, ignore the field of t2 in all compartments and set them to 0 For MTE-SANDIdot model, consider the t2 values in all compartments The fraction vector represents fractions of the soma, neurite and dot with the fraction of extra being 1-sum(fracs)

Combined diffusion-relaxometry models

Microstructure.MTE_SMT — Type

MTE_SMT(axon, extra, fracs, S0norm)

To test multi-TE spherical mean technique for low-b in vivo imaging

Microstructure.MTE_SANDI — Type

MTE_SANDI(soma, neurite, extra, fracs, S0norm)

For MTE-SANDI model, consider the t2 values in all compartments The fraction vector represents fractions of the soma and neurite with the fraction of extra being 1-sum(fracs) S0norm is the relaxation-weighting free signal from all compartments S(b=0,t=0) normalised by S(b=0,t=TEmin)

Prediction of MRI signals

Microstructure.model_signals — Function

model_signals(model,prot[,links])

Predict model signals from BiophysicalModel model and imaging protocol 'prot'. links is a optional argument that specify parameter links in the model