Tissue Compartments

Abstract compartment type

Microstructure.Compartment — Type

Tissue compartments belong to Compartment type. A compartment contains relevant tissue parameters that affect MRI signals

axonal and dendritic compartments

Microstructure.Cylinder — Type

Cylinder(da,dpara,d0,t2)

Cylinder model using Van Gelderen, P. parameters: cylinder diameter 'da' parallel diffusivity 'dpara' intrinsic diffusivity 'd0' T2 relaxation time 't2'

Microstructure.Stick — Type

Stick(dpara,t2)

stick model with zero perpendicular diffusivity parameters: parallel diffusivity 'dpara' T2 relaxation time 't2'

extra-cellular compartment

Microstructure.Zeppelin — Type

Zeppelin(dpara,dperp_frac,t2)

zeppelin/tensor model parameters: parallel diffusivity 'dpara' perpendicular diffusivity represented as a fraction of dpara 'dperp_frac' T2 relaxation time 't2'

cell body compartment

Microstructure.Sphere — Type

Sphere(diff,size,t2)

sphere model (Neuman) parameters: diffusivity within sphere 'diff' sphere radius 'size' T2 relaxation time 't2'

CSF and dot compartment

Microstructure.Iso — Type

Iso(diff,t2)

dot/isotropic tensor parameters: diffusivity 'diff' T2 relaxation time 't2' This compartment can be used to represent CSF (diff = free water) or dot compartment (diff = 0). The latter is for immobile typically seen in ex vivo tissue

Compartment signals

Missing docstring.

Missing docstring for compartment_signals. Check Documenter's build log for details.