I/O functions

The page introduces functions for reading dMRI data and protocols for microstructure imaging.

Read dMRI data and get spherical mean signals & imaging protocols

Microstructure.spherical_mean — Function

spherical_mean(
    image_file::String, 
    normalize::Bool=true, 
    save::Bool=true, 
    acq_files::String...
)

Perform direction average on input DWI images "imagefile" and return an MRI object with spherical mean signal and associated imaging protocol. "imagefile" is the full path of the DWI image file; "normalize" indicates if normalisation to the b0 is applied; "save" indicates if save the smt signals to nifti files on the save path as the input data (the protocol is also saved as a text file). Finall, variable number of "acqfiles" are text files that tell you acquistion parameters of each DWI in the "imagefile'. Accepted file extensions are .bvals/.bvecs/.techo/.tdelta/.tsmalldel.

Besides .bvals/.bvecs for conventional modelling, .tdelta/.tsmalldel files are needed for any models that estimate size, e.g. axon diameter, soma radius. .techo is needed if your data is collected with multiple echo-time and you want to do combined-diffusion relaxometry modelling. The format of a .tdelta/.tsmalldel/.techo file is similar to a .bvals file (a vector with the length equal to the number of DWI volumes). Unit in the .tdelta/.tsmalldel/.techo file is ms.

The dMRI and Protocol type

Microstructure.dMRI — Type

dMRI(nifti::MRI, 
tdelta::Vector{Float64}, 
dsmalldel::Vector{Float64}, 
techo::Vector{Float64}, 
smt::Bool)

Return a dMRI Type object with MRI object 'nifti', and additional volume-wise experimental settings 'tdelta', 'tsmalldel', 'techo', and 'smt' for identifing smt signals.

Microstructure.Protocol — Type

Protocol(
bval::Vector{Float64}
techo::Vector{Float64}
tdelta::Vector{Float64}
tsmalldel::Vector{Float64}
gvec::Vector{Float64}
)

Return a Protocol Type object to hold parameters in acquisition protocol relavent for modelling. Unit convention: most text files use s/mm^2 for b-values and ms for time while they are converted to SI unit in the Protocol. b-values (s/m^2); time (s); size (m); G (T/m)

Write image and save protocol

Microstructure.dmri_write — Function

dmri_write(dmri::dMRI, datapath::String, filename::String)

Write the nifti volume in a dMRI object to nifti file and associated protocol as b-table text files in the given "datapath" and "filename".