I/O functions

dMRI and Protocol type

Microstructure.dMRI — Type

dMRI(nifti,tdelta,dsmalldel,techo,smt)

contain additional volume-wise experimental settings that are needed for advanced modelling expects addtional text files saved besides .bvals/.bvecs expected file extensions: .techo/.tdelta/.tsmalldel

Microstructure.Protocol — Type

Struct to hold parameters in acquisition protocol relavent for modelling Suggest to use SVector when using smt-based modelling and Vector when using orientation models where the number of measurements exceeds 50

unit convention: most text files use s/mm^2 for b-values and ms for time
we use US unit for microsturcture computation
b-values (s/m^2); time (s); size (m); G (T/m) (660 mT/m => 0.66T/m)

Consider to add bvec for orientation modelling or create a new protocol type with gradient orientation info

Read dMRI data and get spherical mean signals & imaging protocols

Missing docstring.

Missing docstring for Spherical_mean. Check Documenter's build log for details.

write dMRI and save protocol

Microstructure.dmri_write — Function

write dmri volume and protocols/b-tables