MCMC Estimators

Define a sampler for your model

Microstructure.Sampler — Type

Sampler for a biophysical model params: parameters (String) to sample in the model priorrange: bounds for each parameter proposal: Distribution to draw pertubations paralinks: linking parameters in the model nsamples: The total number of samples in a MCMC chain; default to 70000 burnin: The number of samples that will be discarded in the beginning of the chain; default to 20000 thinning: The interval to extract unrelated samples in the chain; default to 100 Example Sampler for ExCaliber Sampler( params = ("axon.da","axon.dpara","extra.dperpfrac","fracs") prior_range = ((1.0e-7,1.0e-5),(0.01e-9,0.9e-9),(0.0, 1.0),(0.0,1.0)) proposal = (Normal(0,0.25e-6), Normal(0,0.025e-9), Normal(0,0.05), MvNormal([0.0025 0 0;0 0.0001 0; 0 0 0.0001])) #; (0,0.05),Normal(0,0.01),Normal(0,0.01)]] paralinks = ("axon.d0" => "axon.dpara", "extra.dpara" => "axon.dpara") )

Define a noise model

Missing docstring.

Missing docstring for Noise_model. Check Documenter's build log for details.

run MCMC on your model and data

Microstructure.mcmc! — Function

Run mcmc for a model and sampler

Method 1 generates pertubations within function, creates and returns a dict chain, and modify final model estimates in place. This method is useful in checking a few voxels, e.g. for quality of fitting, chain dignostics and optimizing sampler for models.

Method 2 takes chain and pertubations as input, mutating chain in place which can be used to calculate finial estimates and uncertainties. This method is used for processing larger dataset, e.g. for whole-barin/slices. This method is used together with multi-threads processing that pre-allocate spaces for caching chains, avoiding creating them for each voxel. This method also reuses pertubations for faster speed, as we usually use a very large number of pertubations (e.g. 70000) to densely sample the proposal distributions.

Method 1

julia> mcmc!(estimates,measurements,protocol,sampler,noise_model,rng)

Method 2: 'chain' can be Vector (modify elements) or Dict (push!); need to benchmark time difference

julia> mcmc!(chain,estimates,meas,protocol,sampler,pertubations,noise_model))