Microstructure Models

This page introduces several biophysical models. A biophysical model includes several tissue compartments and other parameters including compartment fractions and signals.

Microstructure.BiophysicalModel — Type

All models in this page belong to the BiophysicalModel Type. You can also build your models with desired combinations of compartments using a similar syntax. In each model, all compartmental parameters can be considered "free parameters" and sampled using MCMC. This is designed to offer maximum flexibility in adjusting model assumptions, but it doesn't guarantee reliable estimation of all parameters. It's common that we need to fix or link some tissue parameters based on our data measurement protocols and our tissue parameters of interest. Parameter fixing and linking can be achieved by settings in MCMC sampler in the estimator module.

dMRI models

WM models

Microstructure.ExCaliber — Type

ExCaliber(
    axon::Cylinder,
    extra::Zeppelin,
    dot::Iso,
    fracs::Vector{Float64}
    )

ExCaliber is a multi-compartment model for estimating axon diameter. It can be used for ex vivo imaging when the diffusivity in the ISO compartment is set to 0 (dot compatment), and for in vivo imaging if the diffusivity of the ISO compartment is set to free water in tissue (CSF compartment).

Reference

Fan, Q., Nummenmaa, A., Witzel, T., Ohringer, N., Tian, Q., Setsompop, K., ... & Huang, S. Y. (2020). Axon diameter index estimation independent of fiber orientation distribution using high-gradient diffusion MRI. Neuroimage, 222, 117197.

Gong, T., Maffei, C., Dann, E., Lee, H.-H., Lee Hansol, Huang, S., Suzanne, H., Yendiki, A., 2024. Imaging the relationship of axon diameter and myelination in macaque and human brain, in: ISMRM.

GM models

Microstructure.SANDI — Type

SANDI(
    soma::Sphere,
    neurite::Stick,
    extra::Iso,
    fracs::Vector{Float64}
    )

The soma and neurite density imaging (SANDI) model uses a sphere compartment to model the cell soma, a stick compartment to model the neurite and an isotropic diffusion compartment for the extra-cellular space; It includes all the tissue parameters in each compartment and a fracs vector representing the fraction of intra-soma signal and intra-neurite signal (the extra-cellular signal fraction is 1-sum(fracs)). For SANDI model, ignore the field of t2 in all compartments and set them to 0.

Reference

Palombo, M., Ianus, A., Guerreri, M., Nunes, D., Alexander, D.C., Shemesh, N., Zhang, H., 2020. SANDI: A compartment-based model for non-invasive apparent soma and neurite imaging by diffusion MRI. Neuroimage 215. https://doi.org/10.1016/j.neuroimage.2020.116835

Microstructure.SANDIdot — Type

SANDIdot(
    soma::Sphere 
    neurite::Stick
    extra::Iso 
    dot::Iso
    fracs::Vector{Float64} 
)

SANDIdot model includes additionally a dot compartment for SANDI model; the dot compartment is considered as immobile water and is more commonly seen in ex vivo imaging. For SANDIdot model, ignore the field of t2 in all compartments and set them to 0. The fraction vector represents fractions of the soma, neurite and dot with the fraction of extra being 1-sum(fracs).

Reference

Alexander, D.C., Hubbard, P.L., Hall, M.G., Moore, E.A., Ptito, M., Parker, G.J.M., Dyrby, T.B., 2010. Orientationally invariant indices of axon diameter and density from diffusion MRI. Neuroimage 52, 1374–1389. https://doi.org/10.1016/j.neuroimage.2010.05.043

Panagiotaki, E., Schneider, T., Siow, B., Hall, M.G., Lythgoe, M.F., Alexander, D.C., 2012. Compartment models of the diffusion MR signal in brain white matter: A taxonomy and comparison. Neuroimage 59, 2241–2254.

Palombo, M., Ianus, A., Guerreri, M., Nunes, D., Alexander, D.C., Shemesh, N., Zhang, H., 2020. SANDI: A compartment-based model for non-invasive apparent soma and neurite imaging by diffusion MRI. Neuroimage 215. https://doi.org/10.1016/j.neuroimage.2020.116835

Combined diffusion-relaxometry models

Microstructure.MTE_SMT — Type

MTE_SMT(
    axon::Stick = Stick()
    extra::Zeppelin = Zeppelin()
    fracs::Float64 = 0.5
    )

This is a model using multi-TE spherical mean technique for lower b-value in vivo imaging. Compartmental T2s are considered. There is not a specific reference for this model yet, but you can refer to previous work related to this topic:

Kaden, E., Kruggel, F., Alexander, D.C., 2016. Quantitative mapping of the per-axon diffusion coefficients in brain white matter. Magn Reson Med 75, 1752–1763. https://doi.org/10.1002/MRM.25734

Kaden, E., Kelm, N. D., Carson, R. P., Does, M. D., & Alexander, D. C. (2016). Multi-compartment microscopic diffusion imaging. NeuroImage, 139, 346-359.

Veraart, J., Novikov, D.S., Fieremans, E., 2017. TE dependent Diffusion Imaging (TEdDI) distinguishes between compartmental T 2 relaxation times. https://doi.org/10.1016/j.neuroimage.2017.09.030

Gong, T., Tong, Q., He, H., Sun, Y., Zhong, J., Zhang, H., 2020. MTE-NODDI: Multi-TE NODDI for disentangling non-T2-weighted signal fractions from compartment-specific T2 relaxation times. Neuroimage 217. https://doi.org/10.1016/j.neuroimage.2020.116906

Microstructure.MTE_SANDI — Type

MTE_SANDI(
    soma::Sphere 
    neurite::Stick
    extra::Iso 
    fracs::Vector{Float64} 
    S0norm::Float64
    )

For Multi-echo-SANDI (MTE-SANDI) model, consider the t2 values in all compartments, and the fractions estimated will be non-T2-weighted compartment fractions in comparison to the model mentioned above. S0norm is the relaxation-weighting free signal from all compartments S(b=0,t=0) normalised by S(b=0,t=TEmin).

Reference

Gong, T., Tax, C.M., Mancini, M., Jones, D.K., Zhang, H., Palombo, M., 2023. Multi-TE SANDI: Quantifying compartmental T2 relaxation times in the grey matter. Toronto.

Prediction of MRI signals

This function implements different methods for different BiophysicalModel types.

Microstructure.model_signals — Function

model_signals(model::BiophysicalModel,prot::Protocol[,links])

Reture predicted model signals from BiophysicalModel model and imaging protocol 'prot'. links is a optional argument that specify parameter links in the model.